I organized the project structure by creating a separate folder for each component with its .jsx file and .module.css file.
