+++ src/components/InputWithLabel/InputWithLabel.module.css
@@ -1,0 +1,7 @@
+input[type="text"] {
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 1rem;
+  margin-right: 10px;
+}
