+++ src/components/TodoList/TodoList.jsx
@@ -1,15 +1,15 @@
-import TodoListItem from "./TodoListItem";
-import "./App.css";
+import TodoListItem from "../TodoListItem/TodoListItem";
+import style from "./TodoList.module.css";
 
 function TodoList({ todoList, onRemoveTodo }) {
   return (
     <>
       <br />
-      <table className="todo-table">
+      <table className={style.table}>
         <thead>
           <tr>
-            <th>Title</th>
-            <th>Actions</th>
+            <th className="todo-table th">Title</th>
+            <th className="todo-table th">Actions</th>
           </tr>
         </thead>
         <tbody>
